🔍 Allow search to filter by category/description
SUPER JANK!

Allows you to filter based on category and description though, which makes it quite a bit more powerful/useful.

There was an issue opened requesting to filter by category and this essentially allows it to do so with the addition of description which is what I want. However, this is quite janky and can be made better through more configuration options.